requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-admin-sortable2==2.2.4
django-colorfield==0.14.0
et_xmlfile==2.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.8.2
openpyxl==3.1.5
pillow==11.1.0
psycopg==3.2.5
psycopg-binary==3.2.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.2
pymdown-extensions==10.16.1
pyOpenSSL==25.0.0
PyYAML==6.0.2
qrcode==8.1
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
whitenoise==6.8.2
zope.interface==7.2
